feat: optimize select flip

+++ docs/.vitepress/cache/deps/chunk-DCJDM2X5.js
@@ -1,4 +1,4 @@
-// node_modules/.pnpm/@vue+shared@3.5.20/node_modules/@vue/shared/dist/shared.esm-bundler.js
+// node_modules/.pnpm/@vue+shared@3.5.21/node_modules/@vue/shared/dist/shared.esm-bundler.js
 function makeMap(str) {
   const map2 = /* @__PURE__ */ Object.create(null)
   for (const key of str.split(',')) map2[key] = 1
@@ -53,9 +53,9 @@ var cacheStringFunction = (fn) => {
     return hit || (cache[str] = fn(str))
   }
 }
-var camelizeRE = /-(\w)/g
+var camelizeRE = /-\w/g
 var camelize = cacheStringFunction((str) => {
-  return str.replace(camelizeRE, (_, c) => (c ? c.toUpperCase() : ''))
+  return str.replace(camelizeRE, (c) => c.slice(1).toUpperCase())
 })
 var hyphenateRE = /\B([A-Z])/g
 var hyphenate = cacheStringFunction((str) => str.replace(hyphenateRE, '-$1').toLowerCase())
@@ -334,7 +334,7 @@ function normalizeCssVarValue(value) {
   return String(value)
 }
 
-// node_modules/.pnpm/@vue+reactivity@3.5.20/node_modules/@vue/reactivity/dist/reactivity.esm-bundler.js
+// node_modules/.pnpm/@vue+reactivity@3.5.21/node_modules/@vue/reactivity/dist/reactivity.esm-bundler.js
 function warn(msg, ...args) {
   console.warn(`[Vue warn] ${msg}`, ...args)
 }
@@ -2040,11 +2040,11 @@ function traverse(value, depth = Infinity, seen) {
   if (depth <= 0 || !isObject(value) || value['__v_skip']) {
     return value
   }
-  seen = seen || /* @__PURE__ */ new Set()
-  if (seen.has(value)) {
+  seen = seen || /* @__PURE__ */ new Map()
+  if ((seen.get(value) || 0) >= depth) {
     return value
   }
-  seen.add(value)
+  seen.set(value, depth)
   depth--
   if (isRef2(value)) {
     traverse(value.value, depth, seen)
@@ -2069,7 +2069,7 @@ function traverse(value, depth = Infinity, seen) {
   return value
 }
 
-// node_modules/.pnpm/@vue+runtime-core@3.5.20/node_modules/@vue/runtime-core/dist/runtime-core.esm-bundler.js
+// node_modules/.pnpm/@vue+runtime-core@3.5.21/node_modules/@vue/runtime-core/dist/runtime-core.esm-bundler.js
 var stack = []
 function pushWarningContext(vnode) {
   stack.push(vnode)
@@ -2503,10 +2503,13 @@ function checkRecursiveUpdates(seen, fn) {
 var isHmrUpdating = false
 var hmrDirtyComponents = /* @__PURE__ */ new Map()
 if (true) {
-  getGlobalThis().__VUE_HMR_RUNTIME__ = {
-    createRecord: tryWrap(createRecord),
-    rerender: tryWrap(rerender),
-    reload: tryWrap(reload)
+  const g = getGlobalThis()
+  if (!g.__VUE_HMR_RUNTIME__) {
+    g.__VUE_HMR_RUNTIME__ = {
+      createRecord: tryWrap(createRecord),
+      rerender: tryWrap(rerender),
+      reload: tryWrap(reload)
+    }
   }
 }
 var map = /* @__PURE__ */ new Map()
@@ -2580,10 +2583,12 @@ function reload(id, newComp) {
       dirtyInstances.delete(instance)
     } else if (instance.parent) {
       queueJob(() => {
-        isHmrUpdating = true
-        instance.parent.update()
-        isHmrUpdating = false
-        dirtyInstances.delete(instance)
+        if (!(instance.job.flags & 8)) {
+          isHmrUpdating = true
+          instance.parent.update()
+          isHmrUpdating = false
+          dirtyInstances.delete(instance)
+        }
       })
     } else if (instance.appContext.reload) {
       instance.appContext.reload()
@@ -3031,23 +3036,26 @@ function hydrateTeleport(
   { o: { nextSibling, parentNode, querySelector, insert, createText } },
   hydrateChildren
 ) {
+  function hydrateDisabledTeleport(node2, vnode2, targetStart, targetAnchor) {
+    vnode2.anchor = hydrateChildren(
+      nextSibling(node2),
+      vnode2,
+      parentNode(node2),
+      parentComponent,
+      parentSuspense,
+      slotScopeIds,
+      optimized
+    )
+    vnode2.targetStart = targetStart
+    vnode2.targetAnchor = targetAnchor
+  }
   const target = (vnode.target = resolveTarget(vnode.props, querySelector))
+  const disabled = isTeleportDisabled(vnode.props)
   if (target) {
-    const disabled = isTeleportDisabled(vnode.props)
     const targetNode = target._lpa || target.firstChild
     if (vnode.shapeFlag & 16) {
       if (disabled) {
-        vnode.anchor = hydrateChildren(
-          nextSibling(node),
-          vnode,
-          parentNode(node),
-          parentComponent,
-          parentSuspense,
-          slotScopeIds,
-          optimized
-        )
-        vnode.targetStart = targetNode
-        vnode.targetAnchor = targetNode && nextSibling(targetNode)
+        hydrateDisabledTeleport(node, vnode, targetNode, targetNode && nextSibling(targetNode))
       } else {
         vnode.anchor = nextSibling(node)
         let targetAnchor = targetNode
@@ -3078,6 +3086,10 @@ function hydrateTeleport(
       }
     }
     updateCssVars(vnode, disabled)
+  } else if (disabled) {
+    if (vnode.shapeFlag & 16) {
+      hydrateDisabledTeleport(node, vnode, node, nextSibling(node))
+    }
   }
   return vnode.anchor && nextSibling(vnode.anchor)
 }
@@ -3191,7 +3203,7 @@ var BaseTransitionImpl = {
       if (
         oldInnerChild &&
         oldInnerChild.type !== Comment &&
-        !isSameVNodeType(innerChild, oldInnerChild) &&
+        !isSameVNodeType(oldInnerChild, innerChild) &&
         recursiveGetSubtree(instance).type !== Comment
       ) {
         let leavingHooks = resolveTransitionHooks(oldInnerChild, rawProps, state, instance)
@@ -3494,6 +3506,7 @@ function useTemplateRef(key) {
   }
   return ret
 }
+var pendingSetRefMap = /* @__PURE__ */ new WeakMap()
 function setRef(rawRef, oldRawRef, parentSuspense, vnode, isUnmount = false) {
   if (isArray(rawRef)) {
     rawRef.forEach((r, i) =>
@@ -3538,6 +3551,7 @@ function setRef(rawRef, oldRawRef, parentSuspense, vnode, isUnmount = false) {
     return !knownTemplateRefs.has(ref22)
   }
   if (oldRef != null && oldRef !== ref2) {
+    invalidatePendingSetRef(oldRawRef)
     if (isString(oldRef)) {
       refs[oldRef] = null
       if (canSetSetupRef(oldRef)) {
@@ -3601,14 +3615,27 @@ function setRef(rawRef, oldRawRef, parentSuspense, vnode, isUnmount = false) {
         }
       }
       if (value) {
-        doSet.id = -1
-        queuePostRenderEffect(doSet, parentSuspense)
+        const job = () => {
+          doSet()
+          pendingSetRefMap.delete(rawRef)
+        }
+        job.id = -1
+        pendingSetRefMap.set(rawRef, job)
+        queuePostRenderEffect(job, parentSuspense)
       } else {
+        invalidatePendingSetRef(rawRef)
         doSet()
       }
     } else if (true) {
       warn$1('Invalid template ref type:', ref2, `(${typeof ref2})`)
     }
+  }
+}
+function invalidatePendingSetRef(rawRef) {
+  const pendingSetRef = pendingSetRefMap.get(rawRef)
+  if (pendingSetRef) {
+    pendingSetRef.flags |= 8
+    pendingSetRefMap.delete(rawRef)
   }
 }
 var hasLoggedMismatchError = false
@@ -8242,8 +8269,9 @@ function emit(instance, event, ...rawArgs) {
     callWithAsyncErrorHandling(onceHandler, instance, 6, args)
   }
 }
+var mixinEmitsCache = /* @__PURE__ */ new WeakMap()
 function normalizeEmitsOptions(comp, appContext, asMixin = false) {
-  const cache = appContext.emitsCache
+  const cache = __VUE_OPTIONS_API__ && asMixin ? mixinEmitsCache : appContext.emitsCache
   const cached = cache.get(comp)
   if (cached !== void 0) {
     return cached
@@ -8696,7 +8724,7 @@ function patchSuspense(
   const { activeBranch, pendingBranch, isInFallback, isHydrating } = suspense
   if (pendingBranch) {
     suspense.pendingBranch = newBranch
-    if (isSameVNodeType(newBranch, pendingBranch)) {
+    if (isSameVNodeType(pendingBranch, newBranch)) {
       patch(
         pendingBranch,
         newBranch,
@@ -8767,7 +8795,7 @@ function patchSuspense(
           )
           setActiveBranch(suspense, newFallback)
         }
-      } else if (activeBranch && isSameVNodeType(newBranch, activeBranch)) {
+      } else if (activeBranch && isSameVNodeType(activeBranch, newBranch)) {
         patch(activeBranch, newBranch, container, anchor, parentComponent, suspense, namespace, slotScopeIds, optimized)
         suspense.resolve(true)
       } else {
@@ -8788,7 +8816,7 @@ function patchSuspense(
       }
     }
   } else {
-    if (activeBranch && isSameVNodeType(newBranch, activeBranch)) {
+    if (activeBranch && isSameVNodeType(activeBranch, newBranch)) {
       patch(activeBranch, newBranch, container, anchor, parentComponent, suspense, namespace, slotScopeIds, optimized)
       setActiveBranch(suspense, newBranch)
     } else {
@@ -9932,7 +9960,7 @@ function getComponentPublicInstance(instance) {
     return instance.proxy
   }
 }
-var classifyRE = /(?:^|[-_])(\w)/g
+var classifyRE = /(?:^|[-_])\w/g
 var classify = (str) => str.replace(classifyRE, (c) => c.toUpperCase()).replace(/[-_]/g, '')
 function getComponentName(Component, includeInferred = true) {
   return isFunction(Component)
@@ -9975,15 +10003,23 @@ var computed2 = (getterOrOptions, debugOptions) => {
   return c
 }
 function h(type, propsOrChildren, children) {
+  const doCreateVNode = (type2, props, children2) => {
+    setBlockTracking(-1)
+    try {
+      return createVNode(type2, props, children2)
+    } finally {
+      setBlockTracking(1)
+    }
+  }
   const l = arguments.length
   if (l === 2) {
     if (isObject(propsOrChildren) && !isArray(propsOrChildren)) {
       if (isVNode(propsOrChildren)) {
-        return createVNode(type, null, [propsOrChildren])
+        return doCreateVNode(type, null, [propsOrChildren])
       }
-      return createVNode(type, propsOrChildren)
+      return doCreateVNode(type, propsOrChildren)
     } else {
-      return createVNode(type, null, propsOrChildren)
+      return doCreateVNode(type, null, propsOrChildren)
     }
   } else {
     if (l > 3) {
@@ -9991,7 +10027,7 @@ function h(type, propsOrChildren, children) {
     } else if (l === 3 && isVNode(children)) {
       children = [children]
     }
-    return createVNode(type, propsOrChildren, children)
+    return doCreateVNode(type, propsOrChildren, children)
   }
 }
 function initCustomFormatter() {
@@ -10182,7 +10218,7 @@ function isMemoSame(cached, memo) {
   }
   return true
 }
-var version = '3.5.20'
+var version = '3.5.21'
 var warn2 = true ? warn$1 : NOOP
 var ErrorTypeStrings = ErrorTypeStrings$1
 var devtools = true ? devtools$1 : void 0
@@ -10204,7 +10240,7 @@ var resolveFilter = null
 var compatUtils = null
 var DeprecationTypes = null
 
-// node_modules/.pnpm/@vue+runtime-dom@3.5.20/node_modules/@vue/runtime-dom/dist/runtime-dom.esm-bundler.js
+// node_modules/.pnpm/@vue+runtime-dom@3.5.21/node_modules/@vue/runtime-dom/dist/runtime-dom.esm-bundler.js
 var policy = void 0
 var tt = typeof window !== 'undefined' && window.trustedTypes
 if (tt) {
@@ -10543,7 +10579,7 @@ function getTransitionInfo(el, expectedType) {
     propCount = type ? (type === TRANSITION ? transitionDurations.length : animationDurations.length) : 0
   }
   const hasTransform =
-    type === TRANSITION && /\b(transform|all)(,|$)/.test(getStyleProperties(`${TRANSITION}Property`).toString())
+    type === TRANSITION && /\b(?:transform|all)(?:,|$)/.test(getStyleProperties(`${TRANSITION}Property`).toString())
   return {
     type,
     timeout,
@@ -10698,7 +10734,7 @@ function setVarsOnNode(el, vars) {
     style[CSS_VAR_TEXT] = cssText
   }
 }
-var displayRE = /(^|;)\s*display\s*:/
+var displayRE = /(?:^|;)\s*display\s*:/
 function patchStyle(el, prev, next) {
   const style = el.style
   const isCssString = isString(next)
@@ -11014,8 +11050,8 @@ function shouldSetAsProp(el, key, value, isSVG) {
 }
 var REMOVAL = {}
 function defineCustomElement(options, extraOptions, _createApp) {
-  const Comp = defineComponent(options, extraOptions)
-  if (isPlainObject(Comp)) extend(Comp, extraOptions)
+  let Comp = defineComponent(options, extraOptions)
+  if (isPlainObject(Comp)) Comp = extend({}, Comp, extraOptions)
   class VueCustomElement extends VueElement {
     constructor(initialProps) {
       super(Comp, initialProps, _createApp)
@@ -11468,7 +11504,7 @@ var TransitionGroupImpl = decorate({
           if (e && e.target !== el) {
             return
           }
-          if (!e || /transform$/.test(e.propertyName)) {
+          if (!e || e.propertyName.endsWith('transform')) {
             el.removeEventListener('transitionend', cb)
             el[moveCbKey] = null
             removeTransitionClass(el, moveClass)
@@ -11995,7 +12031,7 @@ var initDirectivesForSSR = () => {
   }
 }
 
-// node_modules/.pnpm/vue@3.5.20_typescript@5.9.2/node_modules/vue/dist/vue.runtime.esm-bundler.js
+// node_modules/.pnpm/vue@3.5.21_typescript@5.9.2/node_modules/vue/dist/vue.runtime.esm-bundler.js
 function initDev() {
   {
     initCustomFormatter()
@@ -12187,40 +12223,37 @@ export {
 
 @vue/shared/dist/shared.esm-bundler.js:
   (**
-  * @vue/shared v3.5.20
+  * @vue/shared v3.5.21
   * (c) 2018-present Yuxi (Evan) You and Vue contributors
   * @license MIT
   **)
-  (*! #__NO_SIDE_EFFECTS__ *)
 
 @vue/reactivity/dist/reactivity.esm-bundler.js:
   (**
-  * @vue/reactivity v3.5.20
+  * @vue/reactivity v3.5.21
   * (c) 2018-present Yuxi (Evan) You and Vue contributors
   * @license MIT
   **)
 
 @vue/runtime-core/dist/runtime-core.esm-bundler.js:
   (**
-  * @vue/runtime-core v3.5.20
+  * @vue/runtime-core v3.5.21
   * (c) 2018-present Yuxi (Evan) You and Vue contributors
   * @license MIT
   **)
-  (*! #__NO_SIDE_EFFECTS__ *)
 
 @vue/runtime-dom/dist/runtime-dom.esm-bundler.js:
   (**
-  * @vue/runtime-dom v3.5.20
+  * @vue/runtime-dom v3.5.21
   * (c) 2018-present Yuxi (Evan) You and Vue contributors
   * @license MIT
   **)
-  (*! #__NO_SIDE_EFFECTS__ *)
 
 vue/dist/vue.runtime.esm-bundler.js:
   (**
-  * vue v3.5.20
+  * vue v3.5.21
   * (c) 2018-present Yuxi (Evan) You and Vue contributors
   * @license MIT
   **)
 */
-//# sourceMappingURL=chunk-WIXD4IJL.js.map
+//# sourceMappingURL=chunk-DCJDM2X5.js.map
